% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defenses.R
\name{check_cosdata}
\alias{check_cosdata}
\title{Checks for cosdata}
\usage{
check_cosdata(cosdata, opts = coscos::viscos_options())
}
\description{
Perform a set of checks to guarantee tht the provided tht is in the 
`cosdata` format.
}
\keyword{internal}
