% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop}
\alias{chop}
\title{Chop cosdata}
\usage{
chop(cosdata, keys)
}
\arguments{
\item{cosdata}{The strictly defined data format (\code{cosdata}) used 
within \pkg{viscos} (see: \code{\link{cook_cosdata}}).}

\item{keys}{A \code{string} of one or more column-names that designate 
the index-columns for the data separation.}
}
\description{
\code{chop} divides \code{cosdata} into according to chosen column 
identifiers (\code{keys}). The result is a subdivided \code{list}  
which containts the different parts (i.e. rows) of the original data.
}
\examples{
separate the data with regard to the months: 
chopped <- chop(cosdata, key = c("mm"))

separate the data with regard to the year and months: 
chopped <- chop(cosdata, key = c("yyyy","mm"))

}
\seealso{
\code{\link{cook_cosdata}}

Other cosdata manipulators: \code{\link{clump}},
  \code{\link{forge}}
}
\author{
Daniel Klotz
}
