% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{clump}
\alias{clump}
\title{Aggregate cosdata by keys}
\usage{
clump(cosdata, keys = "mm", .funs = mean, opts = coscos::viscos_options())
}
\arguments{
\item{cosdata}{The \emph{cosdata} \code{tibble} (see: \code{\link{cook_cosdata}}).}

\item{keys}{A \code{string}, indicating one or more column-names for data aggregation.}
}
\value{
A aggregated \code{tibble}
}
\description{
Aggregate function, which wraps the \code{dplyr} functions \code{group_by} and \code{summarise_at} so that row-wise aggregation of the \emph{cosdata} \code{tibble} can be made. The operation is defined by a set of functions (given as a \code{list} to the \code{.funs} variable) and the \code{keys} string. The latter speficies the columns, according to whose unique entries the row-wise aggregationis done.
}
\examples{
# aggregate according to month
clump(cosdata, key = "mm", .funs = mean, opts = coscos::viscos_options())

# aggregate according to year and month
clump(cosdata, key = c("yyyy","mm"), .funs = mean, opts = coscos::viscos_options())

}
\seealso{
\code{\link{cook_cosdata}}

Other cosdata manipulators: \code{\link{chop}},
  \code{\link{mould}}
}
\author{
Daniel Klotz
}
