% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{clump_posix}
\alias{clump_posix}
\title{Aggregate posixdate by keys}
\usage{
clump_posix(le_posix, key = "mm", opts = coscos::viscos_options())
}
\description{
Emulates the `clump` function, but for the `posix` column from `cosdata`.
}
