% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{cook_cosdata}
\alias{cook_cosdata}
\title{Cook cosdata}
\usage{
cook_cosdata(raw_data, opts = coscos::viscos_options())
}
\description{
Generates a \code{cosdata tibble} from given \code{raw_data}.
}
\seealso{
Other cooking functions: \code{\link{cook_dates}}
}
