% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{cosdata_as_xts}
\alias{cosdata_as_xts}
\title{Convert cosdata to xts-format}
\usage{
cosdata_as_xts(cosdata, opts = coscos::viscos_options())
}
\value{
xts object of the cosdata data.frame
}
\description{
Converts the cosdata (class: data_frame) into an xts object
}
\keyword{internal}
