% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_measures-GEN.R
\name{d_metrics}
\alias{d_metrics}
\alias{d_nse}
\alias{d_kge}
\alias{d_bias}
\alias{d_pbias}
\alias{d_cor}
\alias{d_mse}
\alias{d_rmse}
\alias{d_inse}
\title{Distance Measures (Metrics)}
\usage{
d_nse(o, s, na.rm = TRUE)

d_kge(o, s, na.rm = TRUE)

d_bias(o, s, na.rm = TRUE)

d_pbias(o, s, na.rm = TRUE)

d_cor(o, s)

d_mse(o, s, na.rm = TRUE)

d_rmse(o, s, na.rm = TRUE)

d_inse(o, s, na.rm = TRUE)
}
\arguments{
\item{o}{The reference data or observations (o_data)}

\item{s}{The created data or the simulations (s_data)}
}
\description{
Different objective Functions, provided by visCOS. A detailed description
of each of the provided objective function is provided in the respective
vignette

Nash-Sutcliffe Efficiency

KGE 2

Percentage Bias

Percentage Bias

Correlation

mean squared error

Root Mean Sqaured Error

Inverted Nash-Sutcliffe Efficiency
}
