% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge.R
\name{forge}
\alias{forge}
\title{Forge cosdata}
\usage{
forge(cosdata, form = "tidy", opts = coscos::viscos_options())
}
\description{
Filter or change `cosdata` into an other format.
}
\seealso{
Other cosdata manipulators: \code{\link{chop}},
  \code{\link{clump}}
}
