% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{libraries}
\alias{libraries}
\title{Loading many libraries at once}
\usage{
libraries(..., verbose = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{\link{require}} and \code{\link{install.packages}}}

\item{x}{character vector. Name(s) of the libraries that are loaded/installed.}
}
\value{
Returns a logical vector that informas if the given package could be 
           loaded
}
\description{
This function tries to load more than one package at once.
If any of these packages is not installed it tries to istall them and load them afterward.
}
\details{
a conveince functions that wraps the `library` function so that 
multiple packages can be loaded with one function call.
}
\examples{
# loading can be done by string or by directly insertin an expression: 
libraries(xts)
libraries(shin)

# load multiple libraries: 
libraries("xts","shiny")
libraries(xts, shiny)

# mixed arguments are also possible:
libraries(magrittr,"shiny","png",jpg) 

}
\seealso{
\code{\link{require}}, \code{\link{library}},
         \code{\link{install.packages}}
}
\author{
Simon Frey
}
