% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2_cosdata.R
\name{map2_cosdata}
\alias{map2_cosdata}
\title{map cosdata}
\usage{
map2_cosdata(cosdata, .f, mapper = purrr::map2)
}
\description{
A pre set \code{map} function for \code{cosdata}. Simultanious application of the 
function \code{.f} to all \eqn{o} and \eqn{s} columns.
}
