% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{mark_periods}
\alias{mark_periods}
\title{Marking-function for defining periods}
\usage{
mark_periods(cosdata, start_month = 10L, end_month = 9L)
}
\arguments{
\item{cosdata}{a data.frame that contains the runoff information.}
}
\value{
`cosdata` with an additonal column with the marked periods.
}
\description{
Compute/Mark the periods within cosdata. The marking uses a monthly
resolution, which are defined by the integers `start_month` and
`end_month`.
}
