% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mould.R
\name{mould}
\alias{mould}
\title{Shape-transformations for cosdata}
\usage{
mould(cosdata, form = "tidy", opts = coscos::viscos_options())
}
\arguments{
\item{cosdata}{The strictly defined data format (\code{cosdata}) used 
within \pkg{visCOS} (see: \code{\link{cook_cosdata}})}

\item{form}{A string that defines into which form the \code{cosdata} is "mouldd"
(see: 'details')}

\item{opts}{The options for \pkg{visCOS} (see: \code{\link{viscos_options}})}
}
\description{
\code{mould} allows to filter, tidy or re-arrange \code{cosdata} into 
new formats.
}
\details{
The mould function transforms  \code{cosdata} into a number of different 
forms. Currently the following forms (strings) can be used:
\itemize{
  \item \emph{tidy} ... Transfroms the \code{cosdata} into a tidy form.
  \item \emph{s_data} ... Filters the \eqn{o} columns out of \code{cosdata}.
  \item \emph{o_data} ... Filters the \eqn{s} columns out of \code{cosdata}.
  \item \emph{s_only} ... Extracts the \eqn{s} columns only.
  \item \emph{o_only} ... Extracts the \eqn{o} columns only.
  \item \emph{shell} ... Returns the \code{cosdata} withouth the 
                           data columns(i.e. the \eqn{o & s} columns).
}
}
\examples{
# tidy: 
mould( coscos::viscos_example() )

# only s-data columns:
mould( coscos::viscos_example(), form = "s_only" )

}
\seealso{
\code{\link{cook_cosdata}}

Other cosdata manipulators: \code{\link{chop}},
  \code{\link{clump}}
}
\author{
Daniel Klotz
}
