% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{objective_functions}
\alias{objective_functions}
\alias{of_nse}
\alias{of_kge}
\alias{of_bias}
\alias{of_pbias}
\alias{of_cor}
\alias{of_mse}
\alias{of_rmse}
\alias{of_inse}
\title{Objective Functions}
\usage{
of_nse(o, s, na.rm = TRUE)

of_kge(o, s, na.rm = TRUE)

of_bias(o, s, na.rm = TRUE)

of_pbias(o, s, na.rm = TRUE)

of_cor(o, s)

of_mse(o, s, na.rm = TRUE)

of_rmse(o, s, na.rm = TRUE)

of_inse(o, s, na.rm = TRUE)
}
\arguments{
\item{o}{The observations or reference data (each row a timesetp and each column a catchment)}

\item{s}{The simulation data (each row a timesetp and each column a catchment)}
}
\description{
Objective Functions (aka \emph{costs} or \emph{performance metrics}) play an important role in hydrology. The following objective functions are the ones mostly used while working with \strong{COSERO}:

Nash-Sutcliffe Efficiency

KGE

Percentage Bias

Percentage Bias

Correlation

Mean Squared Error

Root Mean Squared Error

Inverted Nash-Sutcliffe Efficiency
}
