% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{only_observed_basins}
\alias{only_observed_basins}
\title{remove basins without observations}
\usage{
only_observed_basins(cosdata)
}
\arguments{
\item{cosdata}{A raw cosdata data.frame, which may contains basins without observations. \strong{Note:} It is assumed that all available basins are simulated!}
}
\value{
data.frame without the observation-free basins
}
\description{
Removes basins without observation (-999/NA values) from the provided data.frame
}
\keyword{internal}
