% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{regex1_all_cosdata}
\alias{regex1_all_cosdata}
\title{pre set regular expressions to serch for cosdta column}
\usage{
regex1_all_cosdata(opts)
}
\description{
pre set regular expressions to serch for cosdta column
}
\keyword{internal}
