% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{remove_junk}
\alias{remove_junk}
\title{remove not needed column in cosdata}
\usage{
remove_junk(cosdata, opts = coscos::viscos_options())
}
\arguments{
\item{cosdata}{The cosdata data.frame (see vignette for info)}
}
\value{
data.frame object without the chunk
}
\description{
Removes all columns which are not foreseen (see: viscos_options) from
runoff data.
}
\keyword{internal}
