% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{remove_leading_zeros}
\alias{remove_leading_zeros}
\title{Remove leading zeros from cosdata-names}
\usage{
remove_leading_zeros(cosdata)
}
\description{
Function to remove the leading zeros from the counting part of the \code{cosdata} column-names; e.g. \code{"Qobs_001" -> "qobs_1"}
}
