% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscos_specs.R
\name{viscos_example}
\alias{viscos_example}
\title{Runoff example}
\usage{
viscos_example()
}
\description{
Returns some examplary runoff raw-data that can be transformed into a \emph{cosdata} \code{tibble}, and used to test the functions of \pkg{visCOS}.
}
\examples{
viscos_example() \%>\% head()


}
\seealso{
Other viscos specs: \code{\link{viscos_options}}
}
\author{
Daniel Klotz, Johannes Wesemann, Mathew Herrnegger
}
