% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop}
\alias{chop}
\title{Chop cosdata}
\usage{
chop(cosdata, keys)
}
\arguments{
\item{cosdata}{The \emph{cosdata} \code{tibble} (see: \code{\link{cook_cosdata}}).}

\item{keys}{A \code{string}, indicating one or more column-names for data separation.}
}
\description{
A slicing function, which divides the \emph{cosdata} \code{tibble} into a list of \code{tibbles}. The slicing is done according to one or multiple chosen identifiers columns (referred to as \code{keys}). The envisioned use case is to apply it to subdivide the \emph{cosdata} \code{tibble} according to time dimensiona (e.g. by years). The result is a \code{list} that containts the different parts of the original data.
}
\examples{
# separate the data with regard to months: 
cosdata <- viscos_example()
chopped <- chop(cosdata, keys = c("mm"))
print(head(chopped$slice1_mm))

# separate the data with regard to years and months: 
chopped <- chop(cosdata, keys = c("yyyy","mm"))
print(head(chopped$slice1_yyyy_mm))

}
\seealso{
\code{\link{cook_cosdata}}

Other cosdata manipulators: \code{\link{clump}},
  \code{\link{mould}}
}
\author{
Daniel Klotz
}
