% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{clump}
\alias{clump}
\title{Aggregate cosdata bykeys}
\usage{
clump(cosdata, key = "mm", .funs = mean, opts = coscos::viscos_options())
}
\description{
Provides a wrapper around `dplyr::group_by` and `dplyr::summarise_at` that 
allows for a rowise aggregation of `cosdata` by a set of chosen `key` 
columns.
}
