% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{clump}
\alias{clump}
\title{Aggregate cosdata by keys}
\usage{
clump(cosdata, key = "mm", .funs = mean, opts = coscos::viscos_options())
}
\description{
\code{clump} provides a wrapper around \code{dplyr::group_by} and 
\code{dplyr::summarise_at} that allows for a row-wise aggregation of 
\code{cosdata} by a set of chosen colums (\code{keys})
}
\seealso{
\code{\link{cook_cosdata}}

Other cosdata manipulators: \code{\link{chop}},
  \code{\link{forge}}
}
\author{
Daniel Klotz
}
