% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{clump}
\alias{clump}
\title{Aggregate or clump by keys}
\usage{
clump(cosdata, keys = "mm", .funs = mean, opts = coscos::viscos_options())
}
\arguments{
\item{cosdata}{The \emph{cosdata} \code{tibble} (see: \code{\link{cook_cosdata}}).}

\item{keys}{A \code{string}, indicating one or more column-names for data aggregation.}

\item{.funs}{A aggregation function or list of aggregation functions.}
}
\value{
An aggregated \code{tibble}.
}
\description{
An aggregate function, which wraps the \code{dplyr} functions \code{group_by} and \code{summarise_at} for a row-wise summary of the \emph{cosdata} \code{tibble}. It is defined by the set of functions given by \code{.funs} variable, and the \code{keys} string. The latter specifies the columns, according to whose unique entries the row-wise aggregation is conducted.
}
\examples{
# aggregate according to month
cosdata <- viscos_example()
clump(cosdata, key = "mm", .funs = mean)

# aggregate according to year and month
cosdata <- viscos_example()
clump(cosdata, key = c("yyyy", "mm"), .funs = median)

# aggregate with two functions according to year 
cosdata <- viscos_example()
clump( cosdata, key = "yyyy", .funs = list(a = mean, b = median) )

}
\seealso{
\code{\link{cook_cosdata}}

Other cosdata manipulators: \code{\link{chop}},
  \code{\link{mould}}
}
\author{
Daniel Klotz
}
