% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{clump_posix}
\alias{clump_posix}
\title{Aggregate dates (the posix-colum) by keys}
\usage{
clump_posix(le_posix, keys = "mm")
}
\arguments{
\item{keys}{A \code{string} that contains the short-hand defintions for the aggregation. The used notation is: \code{yyyy} for \emph{years}, \code{mm} for \emph{months} and \code{dd} for \emph{days}.}
}
\value{
A character vector with the aggregated date-entries.
}
\description{
Mirrors the \code{\link{clump}} function, but is used for the \code{posix} column only (see: \code{\link{cook_cosdata}}).
}
\examples{
# clump dates by month
cosdata <- viscos_example() \%>\% cook_cosdata
clump_posix(cosdata$posixdate, keys = "mm")

# clump dates by years
cosdata <- viscos_example() \%>\% cook_cosdata
clump_posix(cosdata$posixdate, keys = "yyyy")

# clump dates by years and months
cosdata <- viscos_example() \%>\% cook_cosdata
clump_posix(cosdata$posixdate, keys = c("yyyy","mm"))

}
