% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{cook_cosdata}
\alias{cook_cosdata}
\title{Generate and check cosdata}
\usage{
cook_cosdata(raw_data, opts = coscos::viscos_options())
}
\arguments{
\item{raw_data}{Some raw-data that can be transformed into the \emph{cosdata} \code{tibble}. The data columns can be specified via the \code{opts} variable (the standard setting uses the \code{visCOS} options, i.e. \code{opts = coscos::viscos_options}).}
}
\value{
A \emph{cosdata} \code{tibble} (see: \code{\link{cook_cosdata}}).
}
\description{
Generates a \emph{cosdata} \code{tibble} from given raw-data (\code{raw_data}).
}
\examples{
# transform raw-data into cosdata:
raw_data <- viscos_example()
head(raw_data)

cos_data <- cook_cosdata(raw_data)
head(cos_data)

}
\seealso{
Other cooking functions: \code{\link{cook_dates}}
}
