% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{libraries}
\alias{libraries}
\title{Load multiple Libraries}
\usage{
libraries(..., verbose = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{\link{library}}.}

\item{vebrose}{A logical vector, indicating if the package start-up messages 
should be supressed (\code{FALSE}) or not (\code{TRUE}).}
}
\value{
Returns a logical vector that informas if the given package could be 
           loaded
}
\description{
\code{libraries} conveniently wraps the \code{\link{base::library}} function so 
that multiple packages can be loaded with a signle one function call.
}
\examples{
# loading can be done by string or by directly insertin an expression: 
libraries(xts)
libraries(shiny)

# load multiple libraries: 
libraries("xts","shiny")
libraries(xts, shiny)

# mixed arguments are also possible:
libraries(magrittr,"shiny","png",stats) 

}
\seealso{
\code{\link{require}}, \code{\link{library}},
         \code{\link{install.packages}}
}
\author{
Simon Frey, Daniel Klotz
}
