% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2_cosdata.R
\name{map2_cosdata}
\alias{map2_cosdata}
\title{map cosdata}
\usage{
map2_cosdata(cosdata, .f, mapper = purrr::map2)
}
\description{
A preset \code{fold} function that wraps around the \link[purrr]{map2}) function from \code{purrr} package, so that the provided function \code{.f} is applied in pairwise order to all \eqn{o} and \eqn{s} columns. The \link[purrr]{map2} function is not fixed; other mapping functions can be used by changing the \code{mapper}.
}
