% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_data.R
\name{mark_periods}
\alias{mark_periods}
\title{Marking-function for defining periods}
\usage{
mark_periods(cosdata, start_month = 10L, end_month = 9L,
  opts = coscos::viscos_options())
}
\arguments{
\item{cosdata}{A \emph{cosdata} \code{tibble} with missing \code{posixdate} column.}
}
\value{
A \emph{cosdata} \code{tibble} with an additional column with the marked periods.
}
\description{
Compure the periods within the \emph{cosdata} \code{tibble}. The function uses a monthly resolution; the bounds of which are defined by the integers \code{start_month} and \code{end_month}.
}
