% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{objective_functions}
\alias{objective_functions}
\alias{of_nse}
\alias{of_kge}
\alias{of_bias}
\alias{of_pbias}
\alias{of_cor}
\alias{of_mse}
\alias{of_rmse}
\alias{of_inse}
\title{Objective Functions}
\usage{
of_nse(o, s, na.rm = TRUE)

of_kge(o, s, na.rm = TRUE)

of_bias(o, s, na.rm = TRUE)

of_pbias(o, s, na.rm = TRUE)

of_cor(o, s)

of_mse(o, s, na.rm = TRUE)

of_rmse(o, s, na.rm = TRUE)

of_inse(o, s, na.rm = TRUE)
}
\arguments{
\item{o}{The reference data or observations (o_data)}

\item{s}{The created data or the simulations (s_data)}
}
\description{
Objective Functions (Performance metrics) play an important role for 
hydrological model calibraiton. Therefore, `coscos` provides a set of innate
objective functions that can be used for model evaluation. The currently 
implemented functions are executed fortran code and the handling is inspired
by the `hydroGOF` package.

Nash-Sutcliffe Efficiency

KGE

Percentage Bias

Percentage Bias

Correlation

Mean Squared Error

Root Mean Squared Error

Inverted Nash-Sutcliffe Efficiency
}
