% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{regex1_all_cosdata}
\alias{regex1_all_cosdata}
\title{Regular Expressions Presets 1}
\usage{
regex1_all_cosdata(opts)
}
\description{
A set of preset regular expressions to search for all cosdata columns.
}
\keyword{internal}
