% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viscos_options.R
\name{viscos_example}
\alias{viscos_example}
\title{Runoff Example}
\usage{
viscos_example()
}
\description{
Returns some examplary runoff data (i.e. raw \code{cosdata}) for 
\pkg{viscos}
}
\seealso{
Other viscos specifica: \code{\link{viscos_options}}
}
